+++ tests/testthat/test-get_elev_raster.R
@@ -11,33 +11,15 @@ sp_sm_prj <- spTransform(sp_sm,CRS(aea_prj))
 test_that("get_elev_raster returns correctly", {
   skip_on_cran()
   skip_on_appveyor()
-  key <- readRDS("key_file.rds")
-  mz_df <- get_elev_raster(locations = pt_df,prj = ll_prj, api_key = key, 
-                           z = 6)
-  mz_sp <- get_elev_raster(locations = sp_sm, api_key = key, 
-                           z = 6)
-  mz_sp_prj <- get_elev_raster(locations = sp_sm_prj, api_key = key, 
-                               z = 6)
-  
-  onetile <- get_elev_raster(locations = sp_sm[1,], api_key = key, 
-                               z = 6)
   
   aws <- get_elev_raster(locations = sp_sm, z = 6, src = "aws")
   aws_prj <- get_elev_raster(locations = sp_sm_prj, z = 6, src = "aws")
   
   #class
-  expect_is(mz_df, "RasterLayer")
-  expect_is(mz_sp, "RasterLayer")
-  expect_is(mz_sp_prj, "RasterLayer")
-  expect_is(onetile,"RasterLayer")
   expect_is(aws,"RasterLayer")
   expect_is(aws_prj,"RasterLayer")
   
   #project
-  expect_equal(proj4string(mz_df),ll_prj)
-  expect_equal(proj4string(mz_sp),ll_prj)
-  expect_equal(proj4string(mz_sp_prj),aea_prj)
-  expect_equal(proj4string(onetile),ll_prj)
   expect_equal(proj4string(aws),ll_prj)
   expect_equal(proj4string(aws_prj),aea_prj)